uievents/order-of-events/mouse-events/click-bubble-manual.html

<!DOCTYPE html>
<html id="html">
  <head>
    <meta charset="utf-8">
    <title>Click-related events should bubble up in the correct order</title>
    <link rel="author" title="Sandra Sun">
    <link rel="help" href="https://w3c.github.io/uievents/#events-clickevent-event-order">
    <meta name="flags" content="interact">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body id="body">
    <ol>
      <li>Double-click into the blue area.</li>
      <li>Hover on the "Done" button.</li>
    </ol>
    <div id="blue" style="width:400px; height:300px; background-color: blue;"></div>
    <button type="button" id="done">Done</button>
</body>
<script>
setup({explicit_timeout: true});
var done = false;
var started = false;
var events = [];
var targets = [];

function record(e) {
  if (done) {
    return;
  }
  events.push(e.type);
  targets.push(e.target.id);
}

var relevantEvents = [
  'mousedown',
  'mouseup',
  'click',
  'dblclick'
];

var relevantTargets = [
  'blue',
  'body',
  'html',
  'document',
  'window'
];

var targets_num = relevantTargets.length;

function finish() {
  done = true;
}

function checkEvents() {
  var order = ['mousedown', 'mouseup', 'click', 'mousedown', 'mouseup',
    'click', 'dblclick'];
  if (events.length != targets_num * order.length
    || targets.length != events.length) {
    return false;
  }

  for (var i = 0; i < order.length; ++i) {
    for (var j = 0; j < targets_num; ++j) {
      if (events[i * targets_num + j] != order[i]
        || targets[i * targets_num + j] != targets[j]) {
        return false;
      }
    }
  }
  return true;
}

window.onload = function () {
  var blue = document.getElementById('blue');
  var inputs = [blue, document.body, document.documentElement, document, window];
  document.id = "document";
  window.id = "window";

  for (var i = 0; i < inputs.length; i++) {
    for (var k = 0; k < relevantEvents.length; k++) {
      inputs[i].addEventListener(relevantEvents[k], record, false);
    }
  }

  async_test(function(t) {
    document.getElementById('done').addEventListener('mouseover', function () {
      finish();
      t.step(function () {
        assert_true(checkEvents(events));
        t.done();
      });
    }, false);
  }, 'Click-related events should bubble up in the correct order');
};
    </script>
</html>
